+++ index.js
@@ -1,3 +1,0 @@
-const pkg = require('./build/build');
-
-exports.Printer = pkg.Printer;