+++ index.ts
@@ -1,3 +1,0 @@
-import {Printer} from "./src/printer";
-
-export {Printer};