+++ lib/host.d.ts
@@ -1,0 +1,9 @@
+export declare function isEdge(): boolean;
+export declare function isIE11(): boolean;
+export declare function isIE10(): boolean;
+export declare function isIE9(): boolean;
+export declare function isIE(): boolean;
+export declare function isChrome(): boolean;
+export declare function isFirefox(): boolean;
+export declare function isSafari(): boolean;
+//# sourceMappingURL=host.d.ts.map