+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+### 2.2.0
+* Change method `splitPage` to `split`
+* Set `Printer` as default export class
+
 ### v1.0.0
 * Initialize package